auction+bidding: use Decimal/Numeric for all money fields
All monetary columns (start_price, minimum_increment, current_price,
last_amount, highest_bid, amount) were Float end to end. The bid
invariant check in Bidding.place_bid() (amount < current_price +
minimum_increment) compared binary floats, so a bid landing exactly
on the minimum increment could be accepted or rejected depending on
rounding artifacts.

- auction_model/auction_read_model/bidding_model/bid_read_model:
  Float -> Numeric(12, 2).
- New Alembic migration (749d800f9644) alters the 8 existing columns
  in place; applied and verified against the local Postgres.
- Auction/Bidding aggregates: money fields typed Decimal instead of
  float. No business logic changed, only the numeric type.
- Pydantic schemas (CreateAuctionSchema, AuctionSchema,
  CreateBidSchema, BidResponseSchema): float -> Decimal.
- Handlers/use case that cast event payload strings back to a number
  (created_auction_handler, scheduled_auction_handler,
  bid_placed_handler x2, auction_started_handler,
  PlaceBidUseCase.execute) now use Decimal instead of float.
- Found and fixed a second bug along the way: BidPlacedEvent put a
  raw Decimal in its payload (unlike every other module, which
  stringifies money in event payloads) which would have crashed the
  websocket broadcast (json.dumps can't serialize Decimal). Bidding.
  place_bid() now stringifies amount like the rest of the codebase.
- Test fixtures/factories that build aggregates or ORM models
  directly (bypassing the Pydantic boundary) now use Decimal too,
  since Decimal and float can't be mixed in arithmetic (+, -) even
  though they can be compared.

API note: money fields now serialize as JSON strings instead of
numbers (Pydantic v2's default for Decimal, avoids reintroducing
float imprecision at the response boundary). Confirmed end to end
with a live request through main.py's lifespan. Frontend will need
to parseFloat/Number() these fields.

72/72 tests passing.

+++ app/modules/auction/application/handlers/bid_placed_handler.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from app.core.events.event_bus_interface import EventBusInterface
 from app.modules.auction.infrastructure.repository.auction_read_repository import (
@@ -31,7 +32,7 @@ class BidPlacedHandler:
 
         current_auction = await self.read_repository.get_by_id(auction_id)
         if current_auction:
-            current_auction.highest_bid = float(event.payload["amount"])
+            current_auction.highest_bid = Decimal(event.payload["amount"])
             current_auction.end_time = auction.end_time
             await self.read_repository.save(current_auction)
