+++ app/modules/bidding/infrastructure/persistence/bid_read_model.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import UUID, Column, DateTime, Float
+from sqlalchemy import UUID, Column, DateTime, Numeric
 
 from app.core.database.base import Base
 
@@ -11,5 +11,5 @@ class BidReadModel(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     auction_id = Column(UUID(as_uuid=True), nullable=False)
     user_id = Column(UUID(as_uuid=True), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     timestamp = Column(DateTime, nullable=False)
