+++ app/modules/bidding/infrastructure/persistence/bidding_model.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import UTC, datetime
 
-from sqlalchemy import Column, DateTime, Float
+from sqlalchemy import Column, DateTime, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 
 from app.core.database.base import Base
@@ -12,10 +12,10 @@ class BiddingModel(Base):
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     auction_id = Column(UUID(as_uuid=True), nullable=False)
-    current_price = Column(Float, nullable=False)
-    minimum_increment = Column(Float, nullable=False)
+    current_price = Column(Numeric(12, 2), nullable=False)
+    minimum_increment = Column(Numeric(12, 2), nullable=False)
     last_user_id = Column(UUID(as_uuid=True), nullable=True)
-    last_amount = Column(Float, nullable=True)
+    last_amount = Column(Numeric(12, 2), nullable=True)
     timestamp = Column(DateTime, default=datetime.now(UTC))
